Modify calculation in setPeriod to be use uint32_t instead of floats

+++ include/EVT/io/PWM.hpp
@@ -25,14 +25,14 @@ public:
      *
      * @param[in] dutyCycle Duty cycle to set the pin to.
      */
-    virtual void setDutyCycle(float dutyCycle) = 0;
+    virtual void setDutyCycle(uint32_t dutyCycle) = 0;
 
     /**
      * Set the period for the PWM in seconds.
      *
-     * @param[in] period The period of the PWM in seconds.
+     * @param[in] period The period of the PWM in microseconds.
      */
-    virtual void setPeriod(float period) = 0;
+    virtual void setPeriod(uint32_t period) = 0;
 
     /**
      * Get the current duty cycle.
